+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-## [1.1.1](https://github.com/rdkcentral/firebolt-cpp-transport/compare/v1.1.0...v1.1.1)
+## [Unreleased](https://github.com/rdkcentral/firebolt-cpp-transport/compare/v1.1.1...main)
 
 ### Changed
-- Events are dispatched in separate threads to avoid blocking the main queue
+- Allow setting log level to `MaxLevel`, fallback to `Debug`
+
+### Fixed
+- Events are dispatched in separate thread to avoid blocking the main queue
+
+## [1.1.1](https://github.com/rdkcentral/firebolt-cpp-transport/compare/v1.1.0...v1.1.1)
 
 ### Fixed
 - Deadlock when unsubscribing from an event inside its callback